36.2.S3 & Private Object via ACLs{Remove that Extra Private Filesystem! }
Remove that Extra Private Filesystem!

Hey! This is awesome! Thanks to the object-by-object permissions super-power of S3, we don't need an extra "private" filesystem at all! We can do some serious cleanup! Start in config/packages/oneup_flysystem.yaml: remove the private_uploads_adapter and filesystem.
Next, in services.yaml, because there's no private_upload_filesystem anymore, remove that bind.
That will break UploaderHelper because we're using that bind on top. But... we don't need it anymore! Remove the $privateFilesystem property and the $privateUploadFilesystem argument.

But, we're still using that property in two places... the first is down in readStream. Now that everything is stored in one filesystem, delete that old code, remove the unused argument and always use $this->filesystem. Reading a stream is the same for public and private files.

Repeat that in deleteFile(): delete the extra logic & argument, and use $this->filesystem always.

Let's see... these two methods are called from ArticleReferenceAdminController. Take off that second argument for readStream().

Then, search for "delete", and remove the second argument from deleteFile() as well.

That felt great! There's one more piece of cleanup we can do, it's optional, but nice. Using the word "public" in the adapter and filesystem isn't accurate anymore! Let's use uploads_adapter and uploads_filesystem.

We reference this in a few spots. In liip_imagine.yaml, take out the public_ in these two spots.

And in services.yaml, update the "bind" in the same way. Hmm, and I think I'll change the argument name it's binding to: just $uploadFilesystem.

That will break UploaderHelper: we need to rename the argument there. But, let's just see what happens if we... "forget" to do that. Refresh the page:
  Unused binding  in S3Client.

This is that generic... and somewhat "inaccurate" error that says that we've configured a bind that's never used!