38.3.Private Downloads & Signed URLs{Creating the Signed URL}
Now that we have this "request" thing... how can we get the URL? Back on their docs, scroll down... here it is: $request->getUri().

When the user hits our endpoint, what we want to do is redirect them to the URL. Do that with return new RedirectResponse(), (string) - they mentioned that in the docs, it turns the URI into a string - then $request->getUri().

Let's try it! Refresh! And... download! Ha! It works! We're loading this directly from S3. This long URL contains a signature that proves to S3 that the request was pre-authenticated and should last for 30 minutes.

But we did lose one thing: our Content-Disposition header! This gave us two nice things: it forced the user to download the file instead of loading it inline, and it controlled the download filename.