39.2.API-Style Uploads{Deserializer & A Model Class}
Deserializer & A Model Class

Love it! Let's get to work. Back in our controller, to see what it looks like, let's make this endpoint capable of handling both ways of uploading files: form-data and JSON.
We can figure out which situation we're in by looking at the Content-Type header. So, if $request->headers->get('Content-Type') === 'application/json', we'll do our new thing, else, run the normal code. And... this is pretty cool... the only part that'll really be different is the $uploadedFile part. Move that into the else.
In the first part of the if, just like a normal API endpoint, we need to decode the JSON request content into something useful. To do that, let's use the serializer! Search for "deser", there it is. Earlier, we used deserialize() to turn the JSON into an ArticleReference object. That worked because the keys in that JSON matched the property names in that class.
But in this case, look at the fields: filename and data. We do have an originalFilename field, and we could rename the filename key to that... but we definitely do not have... and do not want a data property on ArticleReference that's equal to a base64 encoded version of our file. That makes no sense.
This is a classic case where the data of an endpoint doesn't match the structure of our entity. And that's cool! Instead of using the entity, we can create a new model class.
Inside src/, let's create a new Api/ directory - just for organization - and inside, a new class: how about ArticleReferenceUploadApiModel. The whole point of this class is to help us deal with the data for this endpoint. So, its properties should match the data. Add public $filename and public $data.
Yes! Gasp! They're public! Because this class will only be used for this one, narrow, purpose, it's ok to make life a bit easier with public properties. If this makes you want to scream and tackle me, I get it! Just make them private and add the getter & setter methods. That will work perfectly.
While we're here, don't forget about validation: add @Assert\NotBlank above both of these.
We're ready! Back in the controller add a new argument at the end: SerializerInterface $serializer. Then, it's beautiful, really $uploadApiModel = $serializer->deserialize(). This takes three arguments: the raw JSON - $request->getContent() - the type of object it should be turned into - ArticleReferenceUploadApiModel::class - and the input format, json.
We don't need a context this time, because we're not deserializing into an existing object and we don't need to use groups.